BibUpload: ingestion plugins
 * BibUpload now accepts --pre-plugin and --post-plugin
options. A .py file with the same name as the option
passed as argument must be placed in
modules/bibupload/lib/preprocess or
modules/bibupload/lib/postprocess and must include a
function with the same name of the file. The plugin will
receive as argument the full path of the file uploaded.
 * Also BatchUploader accepts these options and forwards
them to BibUpload.